+++ tests/lib/ex/test_move.py
@@ -1,7 +1,5 @@
 from NeoVintageous.tests import unittest
 
-from NeoVintageous.lib.state import State
-
 
 class Test_ex_move_Moving_InNormalMode_SingleLine_DefaultStart(unittest.ViewTestCase):
 
@@ -41,7 +39,7 @@ class Test_ex_move_Moving_InNormalMode_SingleLine_DefaultStart(unittest.ViewTest
         self.write('abc\nxxx\nabc\nabc')
         self.select(4)
 
-        self.view.run_command('ex_move', {'address': '2'})
+        self.view.run_command('ex_move', {'command_line': 'move 2'})
 
         self.assertContent('abc\nxxx\nabc\nabc')
 
@@ -99,28 +97,14 @@ class Test_ex_move_ModeTransition(unittest.ViewTestCase):
         self.write('abc\nxxx\nabc\nabc')
         self.select(4)
 
-        state = State(self.view)
-        state.enter_normal_mode()
+        self.view.run_command('ex_move', {'command_line': 'move 3'})
 
-        self.view.run_command('vi_enter_normal_mode')
-        prev_mode = state.mode
-
-        self.view.run_command('ex_move', {'address': '3'})
-
-        state = State(self.view)
-        new_mode = state.mode
-        self.assertEqual(prev_mode, new_mode)
+        self.assertNormalMode()
 
     def test_from_visual_mode_to_normal_mode(self):
         self.write('abc\nxxx\nabc\nabc')
         self.select((4, 5))
 
-        state = State(self.view)
-        state.enter_visual_mode()
-        prev_mode = state.mode
-
         self.view.run_command('ex_move', {'command_line': 'move3'})
 
-        state = State(self.view)
-        new_mode = state.mode
-        self.assertNotEqual(prev_mode, new_mode)
+        self.assertNormalMode()